refactor: move reusable type declaration to the types directory.
